AoE rewrite and bug fixes
Rewrote aoe checks for cone and line aoes and added minimum distance values
Added height checks for commands
Fixed combo effects repeating for every target hit by AoE attacks
Fixed teleport sometimes not raising (I think)
Fixed gear checks in some command scripts

+++ FFXIVClassic Map Server/actors/chara/ai/StatusEffectContainer.cs
@@ -127,7 +127,7 @@ namespace FFXIVClassic_Map_Server.actors.chara.ai
             return AddStatusEffect(se, owner);
         }
 
-        public bool AddStatusEffect(uint id,  byte tier, UInt64 magnitude)
+        public bool AddStatusEffect(uint id,  byte tier, double magnitude)
         {
             var se = Server.GetWorldManager().GetStatusEffect(id);
 
@@ -137,7 +137,7 @@ namespace FFXIVClassic_Map_Server.actors.chara.ai
             return AddStatusEffect(se, owner);
         }
 
-        public bool AddStatusEffect(uint id, byte tier, UInt64 magnitude, uint duration, int tickMs = 3000)
+        public bool AddStatusEffect(uint id, byte tier, double magnitude, uint duration, int tickMs = 3000)
         {
             var se = Server.GetWorldManager().GetStatusEffect(id);
             if (se != null)
